Se realiza refactor y guardado en localStorage. 1. Se realiza refactor del nombre del dragonball-services a service. 2. Se crea property effect para guardar los characters en el localStorage.
