+++ src/app/pages/dragonball-super/dragonball-super-page.component.ts
@@ -1,4 +1,4 @@
-import { DragonballService } from '../../services/dragonball-services';
+import { DragonballService } from '../../services/dragonball-service';
 import { Component, inject } from '@angular/core';
 import { CharacterListComponent } from '../../components/dragonball/character-list/character-list.component';
 import { CharacterAddComponent } from '../../components/dragonball/character-add/character-add.component';
