+++ src/app/services/dragonball-service.ts
@@ -1,4 +1,4 @@
-import { Injectable, signal } from '@angular/core';
+import { effect, Injectable, signal } from '@angular/core';
 import { Character } from '../components/interfaces/character.interface';
 
 @Injectable({ providedIn: 'root' })
@@ -7,6 +7,12 @@ export class DragonballService {
     { id: 1, name: 'Goku', power: 9001 },
     { id: 2, name: 'Vegeta', power: 8000 },
   ]);
+
+  saveToLocalStorage = effect(() => {
+    localStorage.setItem('characaters', JSON.stringify(this.characters()));
+
+    // console.log(`Character count: ${this.characters().length}`);
+  });
 
   addCharacter(character: Character) {
     this.characters.update((list) => [...list, character]);
